Verify cert is installed properly after writing it
We now verify that the cert is installed by running an `ls`
command after writing it. Ideally we wouldn't need to do that
as the `tar` command will output the cert file name and we would
be able to search for that but for some reason this is not happening.

By verifying that the cert is installed we can say that we return a 200
OK response to the 'inject_client_cert' request only when the cert is
written. This way we won't have a race condition in the client and we
won't respond with a 200 OK if it's not installed. In the latter case we will
log the reason of the failure and raise an error.